\name{RCudaKMeans-package}
\alias{RCudaKMeans-package}
\alias{RCudaKMeans}
\docType{package}
\title{
\packageTitle{RCudaKMeans}
}
\description{
\packageDescription{RCudaKMeans}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCudaKMeans}
\packageIndices{RCudaKMeans}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RCudaKMeans}

Maintainer: \packageMaintainer{RCudaKMeans}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
